ci: Fix coverage command after bugfix in `moon_cove_report`
There was a bug in `moon_cove_report` where the commandline interface
differs with the actual parser implementation. Namely, additional string
arguments were ignored when parsing the args, while in the docs they
should represent input coverage source metadata. The fix made the parser
consistent with the interface, but broke the CI that mistakenly had an
additional `summary` arg in the commandline in here and other projects,
which was ignored in previous versions, but would trigger an error in
the current version.